Graph additions

+++ Assets/Editor Default Resources/Shaders/EditorUITextureShader.shader
@@ -7,7 +7,6 @@
 	SubShader
 	{
 		Tags { "RenderType"="Transparent" }
-
 		Pass
 		{
 			Blend SrcAlpha OneMinusSrcAlpha
@@ -26,7 +25,7 @@
 			struct v2f
 			{
 				float2 uv : TEXCOORD0;
-				float4 color : TEXCOORD1;
+				fixed4 color : TEXCOORD1;
 				float4 vertex : SV_POSITION;
 			};
 
@@ -41,10 +40,12 @@
 				return o;
 			}
 			
-			float4 frag (v2f i) : SV_Target
+			fixed4 frag (v2f i
+			) : SV_Target
 			{
-				float4 col = tex2D(_MainTex, i.uv);
+				fixed4 col = tex2D(_MainTex, i.uv);
 				col *= i.color;
+
 				return col;
 			}
 			ENDCG
